infosys/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
opencv-python==4.9.0.80
mediapipe>=0.10.30
numpy
scipy
openai-whisper
transformers
torch
soundfile
librosa
PyPDF2
python-docx
google-generativeai
ffmpeg-python
